+++ docker/lib/dependabot/docker_compose/file_updater.rb
@@ -7,26 +7,46 @@ module Dependabot
   module DockerCompose
     class FileUpdater < Dependabot::Shared::SharedFileUpdater
       extend T::Sig
+      extend T::Helpers
 
+      YAML_REGEXP = /(docker-)?compose(?>\.[\w-]+)?\.ya?ml/i
       IMAGE_REGEX = /image:\s*/
 
       sig { override.returns(T::Array[Regexp]) }
       def self.updated_files_regex
-        [/(docker-)?compose(?>\.[\w-]+)?\.ya?ml/i]
+        [YAML_REGEXP]
+      end
+
+      sig { override.returns(String) }
+      def file_type
+        "Docker compose"
+      end
+
+      sig { override.returns(Regexp) }
+      def yaml_file_pattern
+        YAML_REGEXP
+      end
+
+      sig { override.returns(Regexp) }
+      def container_image_regex
+        IMAGE_REGEX
+      end
+
+      sig { override.params(escaped_declaration: String).returns(Regexp) }
+      def build_old_declaration_regex(escaped_declaration)
+        %r{#{IMAGE_REGEX}\s+(docker\.io/)?#{escaped_declaration}(?=\s|$)}
       end
 
       sig { override.returns(T::Array[Dependabot::DependencyFile]) }
       def updated_dependency_files
         updated_files = []
-
         dependency_files.each do |file|
           next unless requirement_changed?(file, T.must(dependency))
 
-          updated_files <<
-            updated_file(
-              file: file,
-              content: updated_dockercompose_file_content(file)
-            )
+          updated_files << updated_file(
+            file: file,
+            content: updated_dockerfile_content(file)
+          )
         end
 
         updated_files.reject! { |f| dependency_files.include?(f) }
@@ -35,46 +55,46 @@ module Dependabot
         updated_files
       end
 
-      private
-
-      sig { override.returns(String) }
-      def file_type
-        "docker-compose.yml"
+      sig do
+        override.params(previous_content: String, old_source: T::Hash[Symbol, T.nilable(String)],
+                        new_source: T::Hash[Symbol, T.nilable(String)]).returns(String)
       end
+      def update_digest_and_tag(previous_content, old_source, new_source)
+        old_digest = old_source[:digest]
+        new_digest = new_source[:digest]
 
-      sig { params(file: Dependabot::DependencyFile).returns(String) }
-      def updated_dockercompose_file_content(file)
-        updated_content = update_content(file)
-        raise "Expected content to change!" if updated_content == file.content
+        old_tag = old_source[:tag]
+        new_tag = new_source[:tag]
 
-        updated_content
-      end
-
-      sig { params(file: Dependabot::DependencyFile).returns(String) }
-      def update_content(file)
-        old_source = T.must(previous_sources(file)).first
-        new_source = T.must(sources(file)).first
-
-        content = file.content
-        old_declaration = build_declaration(old_source)
-        new_declaration = build_declaration(new_source)
+        old_declaration =
+          if private_registry_url(old_source)
+            "#{private_registry_url(old_source)}/"
+          else
+            ""
+          end
+        old_declaration += T.must(dependency).name
+        old_declaration +=
+          if specified_with_tag?(old_source)
+            ":#{old_tag}"
+          else
+            ""
+          end
+        old_declaration +=
+          if old_digest&.start_with?("sha256:")
+            "@#{old_digest}"
+          else
+            ""
+          end
 
         escaped_declaration = Regexp.escape(old_declaration)
-        image_regex = %r{^\s*#{IMAGE_REGEX}\s+(docker\.io/)?#{escaped_declaration}(?=\s|$)}
 
-        content.gsub(image_regex) do |old_dec|
-          old_dec.gsub(old_declaration, new_declaration)
+        old_declaration_regex = build_old_declaration_regex(escaped_declaration)
+
+        previous_content.gsub(old_declaration_regex) do |old_dec|
+          old_dec
+            .gsub(":#{old_tag}", ":#{new_tag}")
+            .gsub(old_digest.to_s, new_digest.to_s)
         end
-      end
-
-      sig { params(source: T::Hash[Symbol, T.nilable(String)]).returns(String) }
-      def build_declaration(source)
-        declaration = ""
-        declaration += "#{source[:registry]}/" if source[:registry]
-        declaration += T.must(dependency).name
-        declaration += ":#{source[:tag]}" if source[:tag]
-        declaration += "@sha256:#{source[:digest]}" if source[:digest]
-        declaration
       end
     end
   end
