+++ docker/spec/dependabot/docker_compose/file_parser_spec.rb
@@ -130,21 +130,6 @@ RSpec.describe Dependabot::DockerCompose::FileParser do
           .and_return(status: 200, body: registry_tags)
       end
 
-      context "with no match any tags" do
-        let(:registry_tags) do
-          fixture("docker", "registry_tags", "small_ubuntu.json")
-        end
-
-        before do
-          digest_headers["docker_content_digest"] = "nomatch"
-          ubuntu_url = "https://registry.hub.docker.com/v2/library/ubuntu/"
-          stub_request(:head, /#{Regexp.quote(ubuntu_url)}manifests/)
-            .and_return(status: 200, body: "", headers: digest_headers)
-        end
-
-        its(:length) { is_expected.to eq(0) }
-      end
-
       context "when there is a matching tag" do
         before do
           stub_request(:head, repo_url + "manifests/10.04")
@@ -174,7 +159,7 @@ RSpec.describe Dependabot::DockerCompose::FileParser do
           it "has the right details" do
             expect(dependency).to be_a(Dependabot::Dependency)
             expect(dependency.name).to eq("ubuntu")
-            expect(dependency.version).to eq("12.04.5")
+            expect(dependency.version).to eq("sha256:18305429afa14ea462f810146ba44d4363ae76e4c8dfc38288cf73aa07485005")
             expect(dependency.requirements).to eq(expected_requirements)
           end
         end
@@ -182,26 +167,6 @@ RSpec.describe Dependabot::DockerCompose::FileParser do
         context "with a private registry" do
           let(:composefile_fixture_name) { "private_digest" }
           let(:repo_url) { "https://registry-host.io:5000/v2/myreg/ubuntu/" }
-
-          context "without no/bad authentication credentials" do
-            before do
-              tags_url = repo_url + "tags/list"
-              stub_request(:get, tags_url)
-                .and_return(
-                  status: 401,
-                  body: "",
-                  headers: { "www_authenticate" => "basic 123" }
-                )
-            end
-
-            it "raises a PrivateSourceAuthenticationFailure error" do
-              error_class = Dependabot::PrivateSourceAuthenticationFailure
-              expect { parser.parse }
-                .to raise_error(error_class) do |error|
-                  expect(error.source).to eq("registry-host.io:5000")
-                end
-            end
-          end
 
           context "with good authentication credentials" do
             let(:parser) do
@@ -241,7 +206,7 @@ RSpec.describe Dependabot::DockerCompose::FileParser do
               it "has the right details" do
                 expect(dependency).to be_a(Dependabot::Dependency)
                 expect(dependency.name).to eq("myreg/ubuntu")
-                expect(dependency.version).to eq("12.04.5")
+                expect(dependency.version).to eq("sha256:18305429afa14ea462f810146ba44d4363ae76e4c8dfc38288cf73aa07485005")
                 expect(dependency.requirements).to eq(expected_requirements)
               end
             end
@@ -255,112 +220,6 @@ RSpec.describe Dependabot::DockerCompose::FileParser do
               end
 
               its(:length) { is_expected.to eq(1) }
-            end
-          end
-
-          context "with Amazon ECR" do
-            let(:composefile_fixture_name) { "private_ecr_digest" }
-            let(:repo_url) do
-              "https://695729449481.dkr.ecr.eu-west-2.amazonaws.com/v2/" \
-                "docker-php/"
-            end
-
-            context "without credentials" do
-              before do
-                tags_url = repo_url + "tags/list"
-                stub_request(:get, tags_url)
-                  .and_return(
-                    status: 401,
-                    body: "",
-                    headers: { "www_authenticate" => "basic 123" }
-                  )
-              end
-
-              it "raises a PrivateSourceAuthenticationFailure error" do
-                error_class = Dependabot::PrivateSourceAuthenticationFailure
-                expect { parser.parse }
-                  .to raise_error(error_class) do |error|
-                    expect(error.source)
-                      .to eq("695729449481.dkr.ecr.eu-west-2.amazonaws.com")
-                  end
-              end
-            end
-
-            context "with credentials" do
-              let(:parser) do
-                described_class.new(
-                  dependency_files: files,
-                  credentials: credentials,
-                  source: source
-                )
-              end
-
-              let(:credentials) do
-                [{
-                  "type" => "docker_registry",
-                  "registry" => "695729449481.dkr.ecr.eu-west-2.amazonaws.com",
-                  "username" => "grey",
-                  "password" => "pa55word"
-                }]
-              end
-
-              context "when the private registry invalid" do
-                before do
-                  stub_request(
-                    :post,
-                    "https://api.ecr.eu-west-2.amazonaws.com/"
-                  ).and_return(
-                    status: 403,
-                    body: fixture("docker", "ecr_responses", "invalid_token")
-                  )
-                end
-
-                it "raises a PrivateSourceAuthenticationFailure error" do
-                  error_class = Dependabot::PrivateSourceAuthenticationFailure
-                  expect { parser.parse }
-                    .to raise_error(error_class) do |error|
-                      expect(error.source)
-                        .to eq("695729449481.dkr.ecr.eu-west-2.amazonaws.com")
-                    end
-                end
-              end
-
-              context "when private registry is valid" do
-                subject(:dependency) { dependencies.first }
-
-                before do
-                  stub_request(
-                    :post,
-                    "https://api.ecr.eu-west-2.amazonaws.com/"
-                  ).and_return(
-                    status: 200,
-                    body: fixture("docker", "ecr_responses", "auth_data")
-                  )
-                end
-
-                let(:expected_requirements) do
-                  [{
-                    requirement: nil,
-                    groups: [],
-                    file: "docker-compose.yml",
-                    source: {
-                      registry:
-                        "695729449481.dkr.ecr.eu-west-2.amazonaws.com",
-                      digest: "sha256:18305429afa14ea462f810146ba44d4363ae76" \
-                              "e4c8dfc38288cf73aa07485005"
-                    }
-                  }]
-                end
-
-                its(:length) { is_expected.to eq(1) }
-
-                it "has the right details" do
-                  expect(dependency).to be_a(Dependabot::Dependency)
-                  expect(dependency.name).to eq("docker-php")
-                  expect(dependency.version).to eq("12.04.5")
-                  expect(dependency.requirements).to eq(expected_requirements)
-                end
-              end
             end
           end
         end
